oswan

+++ src/oswan/gpu.h
@@ -28,7 +28,7 @@ extern  unsigned int ws_gpu_unknownPort;
 void ws_gpu_init(void);
 void ws_gpu_done(void);
 void ws_gpu_reset(void);
-void ws_gpu_renderScanline(int16 *framebuffer);
+void ws_gpu_renderScanline(uint8 *framebuffer);
 void ws_gpu_changeVideoMode(uint8 value);
 void ws_gpu_write_byte(uint32_t offset, uint8_t value);
 void ws_gpu_port_write(uint32_t port,uint8_t value);
